Add files via upload
Extremely work in progress addition because PyQt5 is a little limited in resources for this, so it will require ongoing experimentation. This is it in a deliverable state, but until we have reason to use it, it is pretty moot. Will be taking today and tomorrow to give it functionality for showing last time posture was bad.